pc_app/requirements.txt
flask
flask-httpauth
pyside6
pandas
qrcode[pil]
bcrypt
apscheduler
python-dotenv
sendgrid
pyqtgraph